* fixes autoqueue widget button positions * autoqueue: added releaseSpectatorSlot for disconnecting players, just to be sure.
* fixes for map handling when running remote (disable unlink forexample)
* local map browser: shouldn't be able to open it anymore, but in case, added check for running remote + infomessage
* local map browser window -> disable erase maps if you are running remote

+++ Maps/Gui/Windows/AddMaps.php
@@ -82,6 +82,13 @@ class AddMaps extends \ManiaLivePlugins\eXpansion\Gui\Windows\Window
 
 		$login = $this->getRecipient();
 
+
+		if (\ManiaLivePlugins\eXpansion\Helpers\Storage::getInstance()->isRemoteControlled) {
+
+			$this->items[0] = new \ManiaLivePlugins\eXpansion\Adm\Gui\Controls\InfoItem(1, __("File listing disabled since you are running eXpansion remote", $this->getRecipient()), $this->sizeX);
+			$this->pager->addItem($this->items[0]);
+			return;
+		}
 		/** @var \Maniaplanet\DedicatedServer\Structures\Version */
 		$game = $this->connection->getVersion();
 		$path = Helper::getPaths()->getDownloadMapsPath() . $game->titleId . "/*.Map.Gbx";
@@ -99,6 +106,10 @@ class AddMaps extends \ManiaLivePlugins\eXpansion\Gui\Windows\Window
 
 	function deleteMap($login, $filename)
 	{
+		if (\ManiaLivePlugins\eXpansion\Helpers\Storage::getInstance()->isRemoteControlled) {
+			$this->connection->chatSendServerMessage(__("#admin_error#This instance of eXpansion is runnin remotelly! Can't delete file #variable#'%s'", $this->getRecipient(), end($file)));
+			return;
+		}
 		try {
 			unlink($filename);
 			$file = explode("/", $filename);
